requirements.txt
scipy==1.3.1
tensorflow_gpu==1.13.1
imageio==2.5.0
opencv_contrib_python==4.1.1.26
numpy==1.13.3
tqdm==4.36.1
Pillow==7.2.0
tensorflow==2.3.0
